// api/src/app/controllers/acquiring/types/response.types.ts
export interface IAcquiringStart {
  errorCode?: number;
  errorMessage?: string;
  formUrl: string;
  orderId: string;
}

export interface IAcquiringStatus {
  errorCode?: number;
  errorMessage?: string;
  orderNumber: string;
  orderStatus?: number;
  actionCode: number;
  actionCodeDescription: string;
  userMessage?: string;
  amount: number;
  currency?: number;
  date: string;
  depositeddate?: string;
  orderDescription?: string;
  ip?: string;
  authRefNum?: string;
  refundedDate?: number;
  reverseddate?: number;
  paymentWay?: string;
  avsCode?: string;
  refund?: boolean;
  authDateTime?: string;
  terminalId?: string;
  refunds?: IRefunds[];
  attributes?: IAttributes[];
  transactionAttributes?: ITransactionAttributes[];
  merchantOrderParams?: IMerchantOrderParams[];
  cardAuthInfo?: ICardAuthInfo;
  secureAuthInfo?: ISecureAuthInfo;
  bindingInfo?: IBindingInfo;
  paymentAmountInfo?: IPaymentAmountInfo;
  bankInfo?: IBankInfo;
  payerData?: IPayerData;
  dccInfo?: IDccInfo;
  pluginInfo?: IPluginInfo;
}

interface IRefunds {
  date?: string;
  externalRefundId?: string;
  approvalCode?: string;
  actionCode?: number;
  referenceNumber?: string;
  amount?: string;
}

interface IAttributes {
  name?: string;
  value?: string;
}

interface ITransactionAttributes {
  name?: string;
  value?: string;
}

interface IMerchantOrderParams {
  name?: string;
  value?: string;
}

interface ICardAuthInfo {
  maskedPan?: string;
  expiration?: number;
  cardholdername?: string;
  approvalCode?: string;
  paymentSystem: string;
  product: string;
  productCategory: string;
}

interface ISecureAuthInfo {
  eci?: number;
  authTypeIndicator?: string;
  cavv?: string;
  xid?: string;
  threeDSProtocolVersion?: string;
  rreqTransStatus?: string;
  aresTransStatus?: string;
}

interface IBindingInfo {
  clientId?: string;
  bindingId?: string;
  authDateTime?: string;
  authRefNum?: string;
  terminalId?: string;
}

interface IPaymentAmountInfo {
  approvedAmount?: number;
  depositedAmount?: number;
  refundedAmount?: number;
  paymentState?: string;
}

interface IBankInfo {
  bankName?: string;
  bankCountryCode?: string;
  bankCountryName?: string;
}

interface IPayerData {
  email?: string;
  phone?: string;
  postAddress?: string;
}

interface IDccInfo {
  originalCurrencyAlphaCode?: string;
  originalCurrencyNumericCode?: string;
  originalAmount?: string;
  formattedvalue_?: string;
  value?: string;
  convertedCurrencyAlphaCode?: string;
  convertedCurrencyNumericCode?: string;
  convertedAmount?: string;
  formattedvalue?: string;
  conversionRate?: string;
  feePercentage?: string;
  directConversionRate?: string;
  additionalParams?: any;
}

interface IPluginInfo {
  name?: string;
  params?: any;
}